refactor : DBRepository -> DBBoardRepository 클래스명 변경
