+++ src/test/java/chess/domain/chessGame/WhitePawnTest.java
@@ -3,7 +3,7 @@ package chess.domain.chessGame;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.ChessBoard;
-import chess.domain.board.MemoryRepository;
+import chess.domain.board.MemoryBoardRepository;
 import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.position.Column;
@@ -34,7 +34,7 @@ class WhitePawnTest {
     void startPositionPawnWithOnlyAttackablePositions() {
         Position targetPosition = new Position(Row.RANK2, Column.D);
         Color currentTurn = Color.WHITE;
-        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryRepository(
+        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryBoardRepository(
                 Map.of(
                         targetPosition, new Piece(PieceType.WHITE_PAWN, currentTurn),
                         new Position(Row.RANK3, Column.C), new Piece(PieceType.ROOK, currentTurn.opposite()),
@@ -69,7 +69,7 @@ class WhitePawnTest {
     void startPositionPawnWithFreePositions() {
         Position targetPosition = new Position(Row.RANK2, Column.D);
         Color currentTurn = Color.WHITE;
-        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryRepository(
+        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryBoardRepository(
                 Map.of(
                         targetPosition, new Piece(PieceType.WHITE_PAWN, currentTurn),
                         new Position(Row.RANK3, Column.C), new Piece(PieceType.WHITE_PAWN, currentTurn)
@@ -101,7 +101,7 @@ class WhitePawnTest {
     void startPositionPawnWithCantMovePositions() {
         Position targetPosition = new Position(Row.RANK2, Column.D);
         Color currentTurn = Color.WHITE;
-        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryRepository(
+        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryBoardRepository(
                 Map.of(
                         targetPosition, new Piece(PieceType.WHITE_PAWN, currentTurn),
                         new Position(Row.RANK3, Column.C), new Piece(PieceType.WHITE_PAWN, currentTurn),
